Update rgb_colourspace.py
There are only three primaries (3,2) as expected in the RGB colorspace, but in case where large format offset printer gamuts in CMYK colorspace comprising substractive primaries CMY and substractive secondaries RGB, a total of 6 primaries are required to plot the gamut in the CIE xy chromaticity diagram. Hence the change from (3,2) to (6,2), if there is any other way to plot gamuts for CMYRGB primaries please let me know.